Added actuator on the bootstrap.yaml

+++ src/main/java/com/b13/dto/Product.java
@@ -56,5 +56,5 @@ public class Product {
 	Date date_posted;
 	
 	@Version
-	int version;
+	Long version;
 }
